// src/app/Models/GetFlowViewAbscGetApiDataClass.ts
import { DayHourMinuteClass } from "./DayHourMinuteClass";

// Generated by https://quicktype.io

export interface GetFlowViewAbscGetApiDataClass {
    EmpID:             string;
    EmpName:           string;
    DeptName:          string;
    FlowViewAbscDate:  FlowViewAbscDate[];
    UseDayHourMinute:  DayHourMinuteClass;
    Use:               number;
    BaseHour:          number;
    AbscCount:         number;
    Key:               string;
    ProcessFlowID:     number;
    FormName:          string;
    AppEmpID:          string;
    AppEmpName:        string;
    AppDeptName:       string;
    ManageEmpName:     string;
    RealManageEmpName: string;
    State:             string;
    Take:              boolean;
    TransSign:         boolean;
    AbsFlowApps:       any;
    AbscFlowApps:      any;
    CardFlowApps:      any;
    ShiftRoteFlowApps: any;
    Handle:            boolean;
}

export interface FlowViewAbscDate {
    DateTimeB: string;
    DateTimeE: string;
    Use:       number;
}